+++ src/Symfony/Component/PropertyAccess/Tests/DependencyInjection/PropertyAccessBundleExtensionTest.php
@@ -12,6 +12,8 @@
 namespace Symfony\Component\PropertyAccess\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Cache\Adapter\AdapterInterface;
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
@@ -45,12 +47,32 @@ class PropertyAccessBundleExtensionTest extends TestCase
         $container = $this->createContainerFromFile('property_accessor_disabled');
 
         $this->assertFalse($container->has('property_accessor'));
+        $this->assertFalse($container->has('cache.property_access'));
     }
 
-    private function createContainerFromFile(string $file): ContainerBuilder
+    public function testPropertyAccessCache()
+    {
+        $container = $this->createContainerFromFile('property_accessor');
+
+        $cache = $container->getDefinition('cache.property_access');
+        $this->assertSame([PropertyAccessor::class, 'createCache'], $cache->getFactory(), 'PropertyAccessor::createCache() should be used in non-debug mode');
+        $this->assertSame(AdapterInterface::class, $cache->getClass());
+        $this->assertSame([['clearer' => 'cache.system_clearer']], $cache->getTag('cache.pool'));
+    }
+
+    public function testPropertyAccessCacheWithDebug()
+    {
+        $container = $this->createContainerFromFile('property_accessor', true);
+
+        $cache = $container->getDefinition('cache.property_access');
+        $this->assertNull($cache->getFactory());
+        $this->assertSame(ArrayAdapter::class, $cache->getClass(), 'ArrayAdapter should be used in debug mode');
+    }
+
+    private function createContainerFromFile(string $file, bool $debug = false): ContainerBuilder
     {
         $container = new ContainerBuilder();
-        $container->setParameter('kernel.debug', false);
+        $container->setParameter('kernel.debug', $debug);
         $container->registerExtension(new PropertyAccessBundle()->getContainerExtension());
         $container->getCompilerPassConfig()->setBeforeOptimizationPasses([]);
         $container->getCompilerPassConfig()->setOptimizationPasses([]);